requirements.txt
astroid==2.2.5
certifi==2018.11.29
colorama==0.4.1
Django==2.1.5
django-pyodbc-azure==2.1.0.0
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.2
pandas==0.25.1
pylint==2.3.1
pyodbc==4.0.25
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
